clean up test code
I'm going to try one more time and start over again
