--- Game/__init__.py
@@ -1,4 +1,0 @@
-import Game.game
-import Game.loader
-import Game.options
-import Game.map
