+++ Loader.py
@@ -6,15 +6,15 @@ class Loader:
         self.path_to_folder = path_to_folder
         self._image = set()
 
-    def loadimage(self, name, type):
-        if img := self.searchimage(name) is not None:
-            img = image.load(self.path+name+type)
-            self._images.add({name: name, sprite: img})
+    def load_image(self, name, type):
+        if img := self.search_image(name) is not None:
+            img = image.load(self.path_to_folder+name+type)
+            self._image.add({name: name, sprite: img})
             return img
         else:
             return img
 
-    def loadimageArray(self, path_to_animation, type):
+    def load_image_array(self, path_to_animation, type):
         images = []
         counter = 0
         while True:
@@ -25,7 +25,7 @@ class Loader:
                 break
         return images
 
-    def searchimage(self, name):
+    def search_image(self, name):
         for img in self._image:
             if img.name == name:
                 return img.sprite
