--- Objects/__init__.py
@@ -1,6 +1,0 @@
-import Objects.game_objects
-import Objects.player
-import Objects.input
-import Objects.transform
-import Objects.collider
-import Objects.tile
